Resolves circular dependancy - Adds another module for tests
