+++ cordapp-tests/build.gradle
@@ -40,7 +40,8 @@ dependencies {
     testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
 
     // CorDapp dependencies.
-    cordapp project(":cordapp-contracts-states")
+    cordapp project(":cordapp")
+    cordapp project(":services")
 }
 
 task integrationTest(type: Test, dependsOn: []) {
