+++ settings.gradle
@@ -2,4 +2,5 @@ include 'cordapp'
 include 'cordapp-contracts-states'
 include 'clients'
 include 'services'
+include 'cordapp-tests'
